// src/Interfaces/Movies.d.ts
export interface Movies {
	id: string;
	title: string;
	year: string;
	rated: string;
	released: string;
	runtime: string;
	genre: string;
	director: string;
	writer: string;
	actors: string;
	plot: string;
	language: string;
	country: string;
	awards: string;
	poster: string;
	internet_movie_database: string;
	rotten_tomatoes: string;
	metacritic: string;
	metascore: string;
	imdbRating: string;
	imdbVotes: string;
	type: string;
	total_seasons: string;
	dvd: string;
	box_office: string;
	production: string;
	website: string;
	favorite: boolean;
}